Move the controller to its own module
Along with conformance and e2e moved to their own modules, this
greatly simplifies the main module's dependencies, which should make
this less invasive for downstreams.
